rename sidebarNavigation.ts to workspace.ts and pull over workspace/window setup
